--- locale/locale-en.php
@@ -1,84 +1,0 @@
-<?php
-$message = array();
-$message['help'] = 'Help';
-$message['filter'] = 'Filter';
-$message['settings'] = 'Settings';
-$message['language'] = 'Language';
-$message['project'] = 'Project';
-$message['save'] = 'Save';
-$message['settings_show'] = 'Show';
-$message['settings_tool'] = 'Tool';
-$message['settings_wiki'] = 'Wiki';
-$message['settings_bot_edits'] = 'Bot edits';
-$message['settings_anon_edits'] = 'Anon edits';
-$message['settings_minor_edits'] = 'Minor edits';
-$message['settings_redirects'] = 'Redirect pages';
-$message['settings_new_pages'] = 'New pages';
-$message['settings_editor_edits'] = 'Editor edits';
-$message['settings_admin_edits'] = 'Admin edits';
-$message['settings_other_edits'] = 'Other edits';
-$message['send_feedback'] = 'Send feedback';
-$message['time_utc'] = 'Time, in UTC';
-$message['def_i'] = 'v conv'; ## v: verb, conv: conversation
-$message['def_def'] = 'inspect; guard; patrol;';
-$message['main_time_utc'] = 'Time (UTC)';
-$message['main_page'] = 'Page';
-$message['main_user'] = 'User';
-$message['main_info'] = 'Information';
-$message['help_p1'] = 'Raun is a tool to patrol the recent changes of Wikimedia Foundation Project wikis in real time. After a while, Raun will load the recent changes and show it in front of you automatically.';
-$message['help_p2'] = 'You can pause this tool by clicking "Pause". And after that, you can run it again by clicking "Run!".';
-$message['help_p3'] = 'You can filter the edits you want to see under "Filter", and change the wiki project or language, or even the locale of this tool under "Settings".';
-$message['help_p4'] = 'Rows with grey text colour means there are other newer edit(s) on that page.';
-$message['help_legend'] = 'Legend:';
-$message['color'] = 'Color';
-$message['ns'] = 'Namespace';
-$message['ns0'] = 'Article';
-$message['ns1'] = 'Article talk';
-$message['ns2'] = 'User';
-$message['ns3'] = 'User talk';
-$message['ns4'] = 'Project';
-$message['ns5'] = 'Project talk';
-$message['ns6'] = 'File';
-$message['ns7'] = 'File talk';
-$message['ns8'] = 'MediaWiki';
-$message['ns9'] = 'MediaWiki talk';
-$message['ns10'] = 'Template';
-$message['ns11'] = 'Template talk';
-$message['ns12'] = 'Help';
-$message['ns13'] = 'Help talk';
-$message['ns14'] = 'Category';
-$message['ns15'] = 'Category talk';
-$message['ns100'] = 'Portal';
-$message['ns101'] = 'Portal talk';
-$message['close'] = 'Close';
-$message['about'] = 'About';
-$message['about_tool'] = 'Raun is a tool to watch the recent changes of Wikimedia Foundation projects in real time. This tool is inspired from <a href="http://ivan.lanin.org/ronda">Ronda</a>, by Ivan Lanin. This project is started by <a href="http://kenrick95.org/">Kenrick</a> (<a href="http://en.wikipedia.org/wiki/User:Kenrick95">User:Kenrick95</a>).';
-$message['about_cookie'] = 'This site uses <a href="https://en.wikipedia.org/wiki/HTTP_Cookies">cookies</a>. Data stored by the cookie will be deleted automatically if you don\'t visit this tool in 30 days.';
-$message['information'] = 'Information';
-$message['credit'] = 'Credit';
-$message['about_license'] = 'Otherwise stated, this project source code is licensed under MIT License.';
-$message['about_github'] = 'Source code is available on <a href="https://github.com/kenrick95/Raun">github.com/kenrick95/Raun</a>';
-$message['stat'] = 'Statistics';
-$message['stat_articles'] = 'Content pages';
-$message['stat_pages'] = 'Pages';
-$message['stat_files'] = 'Uploaded files';
-$message['stat_edits'] = 'Page edits';
-$message['stat_depth'] = 'Depth';
-$message['stat_users'] = 'Registered users';
-$message['stat_active_users'] = 'Active users';
-$message['stat_admins'] = 'Administrators';
-$message['separator_thousands'] = ',';
-$message['separator_decimals'] = '.';
-$message['settings_pause'] = 'Pause';
-$message['settings_run'] = 'Run!';
-$message['new'] = 'new';
-$message['minor'] = 'minor';
-$message['anon'] = 'anon';
-$message['redirect'] = 'redirect';
-$message['bot'] = 'bot';
-$message['editor'] = 'editor';
-$message['admin'] = 'admin';
-
-
-
-?>
