--- locale/locale-id.php
@@ -1,84 +1,0 @@
-<?php
-$message = array();
-$message['help'] = 'Bantuan';
-$message['filter'] = 'Saringan';
-$message['settings'] = 'Setelan';
-$message['language'] = 'Bahasa';
-$message['project'] = 'Proyek';
-$message['save'] = 'Simpan';
-$message['settings_show'] = 'Tampilkan';
-$message['settings_tool'] = 'Alat';
-$message['settings_wiki'] = 'Wiki';
-$message['settings_bot_edits'] = 'Suntingan bot';
-$message['settings_anon_edits'] = 'Suntingan anon';
-$message['settings_minor_edits'] = 'Suntingan kecil';
-$message['settings_redirects'] = 'Halaman pengalihan';
-$message['settings_new_pages'] = 'Halaman baru';
-$message['settings_editor_edits'] = 'Suntingan editor';
-$message['settings_admin_edits'] = 'Suntingan pengurus';
-$message['settings_other_edits'] = 'Suntingan lainnya';
-$message['send_feedback'] = 'Kirim umpan balik';
-$message['time_utc'] = 'Waktu, dalam UTC';
-$message['def_i'] = 'v cak'; ## v: verba, cak: percakpan
-$message['def_def'] = 'berkeliling; ronda; patroli;';
-$message['main_time_utc'] = 'Waktu (UTC)';
-$message['main_page'] = 'Halaman';
-$message['main_user'] = 'Pengguna';
-$message['main_info'] = 'Keterangan';
-$message['help_p1'] = 'Raun adalah alat untuk memantau perubahan terbaru Wikipedia bahasa Indonesia secara langsung (<i>live</i>). Setiap beberapa saat, Raun akan memuat perubahan terbaru tepat di depan Anda secara otomatis.';
-$message['help_p2'] = 'Anda dapat menjeda (menghentikan sementara) alat ini dengan mengeklik tombol "Jeda". Setelah itu, Anda dapat melanjutkan alat ini dengan mengeklik tombol "Jalan!".';
-$message['help_p3'] = 'Anda dapat menyaring suntingan yang Anda ingin lihat di "Saringan", dan pengaturan proyek wiki, bahasa wiki, dan bahasa alat ini di "Setelan"';
-$message['help_p4'] = 'Baris dengan tulisan berwarna abu-abu berarti terdapat suntingan lain yang lebih baru di halaman tersebut.';
-$message['help_legend'] = 'Legenda:';
-$message['color'] = 'Warna';
-$message['ns'] = 'Ruang nama';
-$message['ns0'] = 'Artikel';
-$message['ns1'] = 'Pembicaraan Artikel';
-$message['ns2'] = 'Pengguna';
-$message['ns3'] = 'Pembicaraan Pengguna';
-$message['ns4'] = 'Proyek';
-$message['ns5'] = 'Pembicaraan Proyek';
-$message['ns6'] = 'Berkas';
-$message['ns7'] = 'Pembicaraan Berkas';
-$message['ns8'] = 'MediaWiki';
-$message['ns9'] = 'Pembicaraan MediaWiki';
-$message['ns10'] = 'Templat';
-$message['ns11'] = 'Pembicaraan Templat';
-$message['ns12'] = 'Bantuan';
-$message['ns13'] = 'Pembicaraan Bantuan';
-$message['ns14'] = 'Kategori';
-$message['ns15'] = 'Pembicaraan Kategori';
-$message['ns100'] = 'Portal';
-$message['ns101'] = 'Pembicaraan Portal';
-$message['close'] = 'Tutup';
-$message['about'] = 'Tentang';
-$message['about_tool'] = 'Raun adalah alat untuk memantau perubahan terbaru Wikipedia bahasa Indonesia secara langsung (<i>live</i>). Alat ini terinspirasi dari <a href="http://ivan.lanin.org/ronda">Ronda</a>, oleh Ivan Lanin. Proyek ini dimulai oleh <a href="http://kenrick95.org/">Kenrick</a> (<a href="http://id.wikipedia.org/wiki/Pengguna:Kenrick95">Pengguna:Kenrick95</a>).';
-$message['about_cookie'] = 'Situs ini menggunakan <a href="https://id.wikipedia.org/wiki/Kuki_HTTP">kuki</a>. Data yang disimpan kuki akan otomatis dihapus jika Anda tidak mengunjungi situs ini dalam 30 hari.';
-$message['information'] = 'Informasi';
-$message['credit'] = 'Kredit';
-$message['about_license'] = 'Di luar lisensi kode dan ide yang telah disebutkan, kode proyek ini dilisensikan dengan lisensi MIT';
-$message['about_github'] = 'Kode tersedia di <a href="https://github.com/kenrick95/Raun">github.com/kenrick95/Raun</a>';
-$message['stat'] = 'Statistik';
-$message['stat_articles'] = 'Halaman konten';
-$message['stat_pages'] = 'Jumlah halaman';
-$message['stat_files'] = 'Jumlah berkas';
-$message['stat_edits'] = 'Jumlah suntingan';
-$message['stat_depth'] = 'Kedalaman';
-$message['stat_users'] = 'Jumlah pengguna';
-$message['stat_active_users'] = 'Pengguna aktif';
-$message['stat_admins'] = 'Jumlah pengurus';
-$message['separator_thousands'] = '.';
-$message['separator_decimals'] = ',';
-$message['settings_pause'] = 'Jeda';
-$message['settings_run'] = 'Jalan!';
-$message['new'] = 'baru';
-$message['minor'] = 'kecil';
-$message['anon'] = 'anon';
-$message['redirect'] = 'alih';
-$message['bot'] = 'bot';
-$message['editor'] = 'editor';
-$message['admin'] = 'admin';
-
-
-
-?>
