--- locale/locale-ta.php
@@ -1,82 +1,0 @@
-<?php
-$message = array();
-$message['help'] = 'உதவி';
-$message['filter'] = 'Filter';
-$message['settings'] = 'அமைப்புகள்';
-$message['language'] = 'மொழி';
-$message['project'] = 'திட்டம்';
-$message['save'] = 'சேமி';
-$message['settings_show'] = 'காட்டு';
-$message['settings_tool'] = 'கருவி';
-$message['settings_wiki'] = 'விக்கி';
-$message['settings_bot_edits'] = 'தானியங்கி தொகுப்புகள்';
-$message['settings_anon_edits'] = 'ஐ.பி தொகுப்புகள்';
-$message['settings_minor_edits'] = 'சிறு தொகுப்புகள்';
-$message['settings_redirects'] = 'வழிமாற்று பக்கங்கள்';
-$message['settings_new_pages'] = 'புதிய பக்கங்கள்';
-$message['settings_editor_edits'] = 'பயனர் தொகுப்புகள்';
-$message['settings_admin_edits'] = 'நிருவாகி தொகுப்புகள்';
-$message['settings_other_edits'] = 'Other edits';
-$message['send_feedback'] = 'Send feedback';
-$message['time_utc'] = 'நேரம், ஒ.ச.நே';
-$message['def_i'] = 'வினைச்சொல்'; ## வி: வினைச்சொல், conv: உரையாடல்
-$message['def_def'] = 'ஆய்விடு; பாதுகாத்திடு; ரோந்திடு;';
-$message['main_time_utc'] = 'நேரம் (ஒ.ச.நே)';
-$message['main_page'] = 'பக்கம்';
-$message['main_user'] = 'பயனர்';
-$message['main_info'] = 'தகவல்';
-$message['help_p1'] = 'Raun கருவி விக்கிமீடிய நிறுவனத்தினது விக்கிகளின் அண்மைய மாற்றங்களை நிகழ்நேரத்தில் ரோந்திட பயன்படும் கருவியாகும். தானாகவே இக்கருவி சிறுநேர இடைவேளியில் அண்மைய மாற்றங்களை புதுப்பித்து உங்களுக்கு காட்டும்.';
-$message['help_p2'] = 'நீங்கள் தற்காலிகமாக கருவியினை நிறுத்திட "நிறுத்து" பொத்தானை சொடுக்கவும். அதன்பிறகு மீண்டும் கருவி இயங்க "இயங்கு!" பொத்தானை சொடுக்கவும்.';
-// $message['help_p3'] = 'அதைத் தவிர, நீங்கள் தொகுப்புகளை வடிகட்டலாம்.';
-$message['help_p3'] = 'You can filter the edits you want to see under "Filter", and change the wiki project or language, or even the locale of this tool under "Settings".';
-$message['help_p4'] = 'Rows with grey text colour means there are other newer edit(s) on that page.';
-$message['help_legend'] = 'உரைகூற்று:';
-$message['color'] = 'நிறம்';
-$message['ns'] = 'பெயர்வெளி';
-$message['ns0'] = 'கட்டுரை';
-$message['ns1'] = 'கட்டுர பேச்சு';
-$message['ns2'] = 'பயனர்';
-$message['ns3'] = 'பயனர் பேச்சு';
-$message['ns4'] = 'திட்டம்';
-$message['ns5'] = 'திட்டம் பேச்சு';
-$message['ns6'] = 'படிமம்';
-$message['ns7'] = 'படிமம் உதவி';
-$message['ns8'] = 'மீடியாவிக்கி';
-$message['ns9'] = 'மீடியாவிக்கி பேச்சு';
-$message['ns10'] = 'வார்ப்புரு';
-$message['ns11'] = 'வார்ப்புரு பேச்சு';
-$message['ns12'] = 'உதவி';
-$message['ns13'] = 'உதவி பேச்சு';
-$message['ns14'] = 'பகுப்பு';
-$message['ns15'] = 'பகுப்பு பேச்சு';
-$message['ns100'] = 'வலைவாசல்';
-$message['ns101'] = 'வலைவாசல் பேச்சு';
-$message['close'] = 'மூடு';
-$message['about'] = 'இதுபற்றி';
-$message['about_tool'] = 'Raun கருவி விக்கிமீடிய நிறுவனத்தினது விக்கிகளின் அண்மைய மாற்றங்களை நிகழ்நேரத்தில் ரோந்திட பயன்படும் கருவியாகும். இக்கருவி ஐவன் லானினின்<a href="http://ivan.lanin.org/ronda">ரோன்டா</a> கருவியினைத் தழுவி எழுதப்பட்டது. இந்த திட்டம் <a href="http://kenrick95.org/">கென்ரிக்</a> (<a href="http://en.wikipedia.org/wiki/User:Kenrick95">User:Kenrick95</a>) ஆல் தொடங்கப்பட்டது.';
-$message['about_cookie'] = 'இந்தப் பக்கம் <a href="https://en.wikipedia.org/wiki/HTTP_Cookies">நினைவிகளைப்</a> பயன்படுத்துகிறது. இப்பக்கத்தினை நீங்கள் பயன்படுத்தவில்லை என்றால் சேமிக்கப்பட்டத் தகவல்கள் 30 நாட்களில் நீக்கப்படும்.';
-$message['information'] = 'தகவல்';
-$message['credit'] = 'நன்றிகள்';
-$message['about_license'] = 'இக்கருவியின் மூலம் MIT உரிமத்தின் கீழ் தரப்பட்டுள்ளது.';
-$message['about_github'] = 'இக்கருவியின் மூலம்: <a href="https://github.com/kenrick95/Raun">github.com/kenrick95/Raun</a>';
-$message['stat'] = 'புள்ளிவிவரங்கள்';
-$message['stat_articles'] = 'உள்ளடக்கம் கொண்ட பக்கங்கள்';
-$message['stat_pages'] = 'பக்கங்கள்';
-$message['stat_files'] = 'பதிவேற்றப்பட்ட கோப்புகள்';
-$message['stat_edits'] = 'பக்க திருத்தங்கள்';
-$message['stat_depth'] = 'ஆழம்';
-$message['stat_users'] = 'பதிவுசெய்த பயனர்கள்';
-$message['stat_active_users'] = 'தொடர் பங்களிப்பாளர்கள்';
-$message['stat_admins'] = 'நிருவாகிகள்';
-$message['separator_thousands'] = ',';
-$message['separator_decimals'] = '.';
-$message['settings_pause'] = 'நிறுத்து';
-$message['settings_run'] = 'இயங்கு!';
-$message['new'] = 'புது';
-$message['minor'] = 'சிறு';
-$message['anon'] = 'அனானி';
-$message['redirect'] = 'வழிமாற்று';
-$message['bot'] = 'தானியங்கி';
-$message['editor'] = 'பயனர்';
-$message['admin'] = 'நிருவாகி';
-?>
